backend/requirements.txt
fastapi
uvicorn[standard]
asyncpg
pydantic
pydantic-settings
python-jose[cryptography]
bcrypt
